requirements.txt
python-dotenv
rich
requests
flask
